requirements.txt
azure-common==1.1.25
azure-core==1.7.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-nspkg==3.0.2
azure-storage-blob==12.3.2
beautifulsoup4==4.8.2
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.1
chardet==3.0.4
click==7.1.1
cryptography==3.0
Flask==1.1.2
Flask-WTF==0.14.3
google-api-core==1.22.0
google-auth==1.20.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
msrest==0.6.18
numpy==1.18.1
oauthlib==3.1.0
opencensus==0.7.10
opencensus-context==0.1.1
opencensus-ext-azure==1.0.4
opencensus-ext-logging==0.1.0
opencensus-ext-requests==0.7.3
pandas==1.0.1
protobuf==3.12.4
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
wtf==0.1
WTForms==2.2.1